move-only: move SourceLocation to separate header
This is preparatory work for a future commit so SourceLocation can
be used without requiring logging.h.
